refactor: use incomingSubmission for old endpoint
